% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mDLNM-package.R
\name{mDLNM}
\alias{mDLNM}
\title{mDLNM: Meta-package for distributed lag non-linear models}
\description{
A meta-package that supports four types of DLNMs:
(1) single-index adaptive cumulative exposure DLNM,
(2) additive adaptive cumulative exposure DLNM,
(3) single-index distributed response function DLNM, and
(4) additive distributed response function DLNM.
}
