% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacking_weights.R
\name{stacking_weights}
\alias{stacking_weights}
\title{Compute stacking weights}
\usage{
stacking_weights(lpd_point, optim_method = "BFGS", optim_control = list())
}
\description{
This function is from loo package: https://github.com/stan-dev/loo/blob/master/R/loo_model_weights.R
}
